size_t -> uint64_t for index type

+++ include/oneapi/dpl/pstl/hetero/dpcpp/parallel_backend_sycl_merge_sort.h
@@ -424,7 +424,7 @@ __parallel_sort_impl(oneapi::dpl::__internal::__device_backend_tag, _ExecutionPo
     }
     else
     {
-        return __submit_selecting_leaf<std::size_t>(std::forward<_ExecutionPolicy>(__exec), std::forward<_Range>(__rng),
+        return __submit_selecting_leaf<std::uint64_t>(std::forward<_ExecutionPolicy>(__exec), std::forward<_Range>(__rng),
                                                     __comp);
     }
 }
